add Repository

+++ src/main/java/by/itstep/internetMarket/service/UserService.java
@@ -8,11 +8,12 @@ import java.util.List;
 public interface UserService {
     void createUser(User user);
 
-    void deleteUser(int id);
+    void deleteUser(Long id);
 
     void updateUser(User user);
 
     List listUsers();
 
     User getUser(String name);
+
 }
